+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.0.1
+
+- Fixed dead enemies being included on the entrance nearby activity check.
+- Added some additional configuration options.
+
 ### 1.0.0
 
 - Initial Release